acpi: fix int-to-pointer-cast warning
A recent patch missed part of the type cast, causing a warning
in some 32-bit builds:

arch/x86/boot/compressed/acpi.c: In function 'get_acpi_srat_table':
arch/x86/boot/compressed/acpi.c:316:9: error: cast to pointer from integer of different size [-Werror=int-to-pointer-cast]

Add the same cast that is used for the normal case.

Fixes: 41fa1ee9c6d6 ("acpi: Ignore acpi_rsdp kernel param when the kernel has been locked down")

+++ arch/x86/boot/compressed/acpi.c
@@ -313,7 +313,7 @@ static unsigned long get_acpi_srat_table(void)
 	 * stash this in boot params because the kernel itself may have
 	 * different ideas about whether to trust a command-line parameter.
 	 */
-	rsdp = (struct acpi_table_rsdp *)get_cmdline_acpi_rsdp();
+	rsdp = (struct acpi_table_rsdp *)(long)get_cmdline_acpi_rsdp();
 
 	if (!rsdp)
 		rsdp = (struct acpi_table_rsdp *)(long)
